backend/requirements.txt
django==2.2.10
djangorestframework
django-cors-headers
django-rest-auth
django-allauth
pillow
django-private-storage
python-magic